requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.24.3
mlxtend==0.23.0
mysql-connector-python==8.2.0
scikit-learn==1.3.0
pydantic==2.5.0
python-dotenv==1.0.0